Extraigo constantes a archivo
