--- frontend-dolce-margarita/src/componentes/home/moldes.js
@@ -1,5 +1,0 @@
-export const moldes = [
-    {nombreCategoria: 'figuras', tituloCarta: 'Figuras'},
-    {nombreCategoria: 'bombones', tituloCarta: 'Bombonería'},
-    {nombreCategoria: 'huevos', tituloCarta: 'Huevos'}
-]
